# api/openapi.yaml
openapi: 3.0.0
info:
  title: BulkTrack API
  version: 1.0.0
  description: API definition for the BulkTrack application.
servers:
  - url: https://api.bulktrack.app/v1
    description: Production server
  - url: http://localhost:8787/v1
    description: Local development server
paths:
...
  /v1/menus:
    get:
      summary: List user menus
      operationId: listMenus
      tags:
        - Menus
      description: Retrieves a list of menus belonging to the authenticated user/device, ordered by sort_order.
      responses:
        '200':
          description: A list of menus.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuDTO'
        '401':
          description: Unauthorized (Invalid or missing Authorization header).
        '500':
          description: Internal Server Error.
      security:
        - bearerAuth: [] # Explicitly require bearer auth for this endpoint

    post:
      summary: Create a new menu
      operationId: createMenu
      tags:
        - Menus
      description: Creates a new menu item for the authenticated user/device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuRequest'
      responses:
        '201':
          description: Menu created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuDTO'
        '400':
          description: Bad Request (e.g., validation error, missing required fields like name).
        '401':
          description: Unauthorized (Invalid or missing Authorization header).
        '500':
          description: Internal Server Error.
      security:
        - bearerAuth: [] # Explicitly require bearer auth for this endpoint

  # --- Workout Endpoints --- #
...
components:
  schemas:
...
    MenuDTO:
      type: object
      description: Represents a menu item.
      properties:
        id:
          type: string
          format: uuid
          description: Menu ID.
        name:
          type: string
          description: Name of the menu.
        description:
          type: string
          nullable: true
          description: Optional description for the menu.
        sort_order:
          type: integer
          description: Order in which the menu should appear.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the menu was created.
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the menu was last updated.
      required:
        - id
        - name
        - sort_order
        - created_at
        - updated_at
    CreateMenuRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the menu item.
          minLength: 1
        description:
          type: string
          nullable: true
          description: Optional description for the menu item.
        sort_order:
          type: integer
          description: Order for sorting menus. Lower values appear first. Defaults to 0 if not provided by client, but API requires it.
      required:
        - name
        - sort_order

    # --- New Workout Schemas for POST /v1/workouts --- #
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] # Default security for endpoints not explicitly marked otherwise
...